style review: PEP 8 changes, removed commented-out code, cite sources
Some of the formatting changes were performed automatically by the
`autopep8` and `isort` tools.

Just because the author of the "Bag of Words Meets Bags of Popcorn"
tutorial (now cited by the actual URL instead of as being "off Kaggle")
bizarrely mixed camelCase and snake_case doesn't mean we should copy
their mistake verbatim, and snake_case is the way we do it in Python.